Move the class RoomDirectoryData form the SDK to the app
