+++ vector/src/main/java/im/vector/app/features/roomdirectory/RoomDirectoryData.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Matrix.org Foundation C.I.C.
+ * Copyright (c) 2021 New Vector Ltd
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.matrix.android.sdk.api.session.room.model.thirdparty
+package im.vector.app.features.roomdirectory
 
 /**
  * This class describes a rooms directory server.
