+++ vector/src/main/java/im/vector/app/features/roomdirectory/picker/RoomDirectoryListCreator.kt
@@ -18,8 +18,8 @@ package im.vector.app.features.roomdirectory.picker
 
 import im.vector.app.R
 import im.vector.app.core.resources.StringArrayProvider
+import im.vector.app.features.roomdirectory.RoomDirectoryData
 import org.matrix.android.sdk.api.session.Session
-import org.matrix.android.sdk.api.session.room.model.thirdparty.RoomDirectoryData
 import org.matrix.android.sdk.api.session.room.model.thirdparty.ThirdPartyProtocol
 import javax.inject.Inject
